+++ silverneedle/lib/Characters/Attacks/IAttackStatistic.cs
@@ -8,8 +8,14 @@ namespace SilverNeedle.Characters.Attacks
     using SilverNeedle.Dice;
     public interface IAttackStatistic
     {
+        string Name { get; }
         BasicStat AttackBonus { get; }
         BasicStat DamageModifier { get; }
         Cup Damage { get; }
+        int NumberOfAttacks { get; }
+        AttackTypes AttackType { get; }
+        BasicStat CriticalModifier { get; }
+        int CriticalThreat { get; }
+        int SaveDC { get; }
     }
 }