+++ CHANGELOG.md
@@ -3,8 +3,9 @@ All notable changes will be tracked in this file. Most likely.
 
 ## [Unreleased Changes]
 ### Added
-- Druid class is now supported (minus animal companions >0.0<
+- Druid class is now supported (minus animal companions >^.^< )
 - Add Spell Resistance statistic
+- Multiple Attacks are awarded for high attack bonuses now
 
 ### Changed
 - Levels were reusing abilities which caused weird bugs, now each level up will create new instances for each character